+++ src/test/resources/mongo-db/mongo-init.js
@@ -1,0 +1,2 @@
+db.createCollection('photobookMeta');
+db.createCollection('photobookImages');