+++ CHANGELOG.md
@@ -10,11 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - dockerfile for native image
+- test containers configuration
 
 ### Changed
 
 - registered classes for reflection
-- mongo db url set to ${MONGODB_URL:mongodb://root:example@localhost:27017/test_db}
+- mongo db url set to ${MONGODB_URL:mongodb://localhost:27017/photobook_demo}
 - default port set to 8080
 - fj-core set to 8.6.2
 - java version 21